chore: re-organise files and home feed order
